+++ Packages/UGF.Initialize/Runtime/InitializeHandler.cs
@@ -1,0 +1,8 @@
+namespace UGF.Initialize.Runtime
+{
+    /// <summary>
+    /// Represents a delegate for initialize handler.
+    /// </summary>
+    /// <param name="handler">The initialize handler.</param>
+    public delegate void InitializeHandler(IInitialize handler);
+}